Add custom.js , where you can extend your pipeline with more transformations
